Rename the file names about WallModel
