+++ examples/ALE_ns_rotate/preprocess.cpp
@@ -15,13 +15,13 @@
 #include "NodalBC_3D_inflow.hpp"
 #include "NodalBC_3D_rotated.hpp"
 #include "ElemBC_3D_outflow.hpp"
-#include "ElemBC_3D_turbulence_wall_model.hpp"
+#include "ElemBC_3D_WallModel.hpp"
 #include "Interface_pair.hpp"
 #include "NBC_Partition.hpp"
 #include "NBC_Partition_inflow.hpp"
 #include "NBC_Partition_rotated.hpp"
 #include "EBC_Partition_outflow.hpp"
-#include "EBC_Partition_turbulence_wall_model.hpp"
+#include "EBC_Partition_WallModel.hpp"
 #include "Interface_Partition.hpp"
 #include "yaml-cpp/yaml.h"
